+++ webserver/src/activity-logger/logger-middlewares/annotation-comment-logger.js
