+++ webserver/src/activity-logger/logger-middlewares/authentication-logger.js
