+++ webserver/src/activity-logger/logger-middlewares/course-logger.js
