+++ webserver/src/activity-logger/logger-middlewares/material-logger.js
