+++ webserver/src/activity-logger/logger-middlewares/topic-logger.js
